Use tempfile.mkstemp instead of NamedTemporaryFile in convert engine
According to https://code.djangoproject.com/wiki/NamedTemporaryFile
django.core.files.temp.NamedTemporaryFile does not support the argument
delete and under Windows is not for the case that a file is opened
in multiple processes which is done in the convert engine because
of the calls via subprocess. So under Windows using NamedTemporaryFile
either from django or Python stdlib will result in non function
convert engine.